requirements.txt
annotated-types==0.7.0
baml-py==0.218.1
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
setuptools==80.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
wheel==0.46.3
